source category provide height index
