+++ noisemodelling-emission/src/test/java/org/noise_planet/noisemodelling/emission/EvaluateRailWaySourceCNOSSOSTest.java
@@ -307,113 +307,6 @@ public class EvaluateRailWaySourceCNOSSOSTest {
 
     }
 
-    private void generateLine(StringBuilder sb, double startX, double startY, double stopX, double stopY) {
-        sb.append(String.format(Locale.ROOT, "<line x1=\"%f\" y1=\"%f\" x2=\"%f\" y2=\"%f\"" +
-                " stroke=\"black\" stroke-width=\"1\" />\n",startX, startY, stopX, stopY));
-    }
-
-    private void generateDashedLine(StringBuilder sb, double startX, double startY, double stopX, double stopY) {
-        sb.append(String.format(Locale.ROOT, "<line x1=\"%f\" y1=\"%f\" x2=\"%f\" y2=\"%f\"" +
-                " stroke=\"black\" stroke-width=\"1\"   stroke-dasharray=\"5,5\" />\n",startX, startY, stopX, stopY));
-    }
-
-    private void generateText(StringBuilder sb, double startX, double startY, int fontSize, String text, String verticalAlignement) {
-        sb.append(String.format(Locale.ROOT, "<text x=\"%f\" y=\"%f\" font-family=\"Verdana\" font-size=\"%d\"  text-anchor=\"middle\" dominant-baseline=\"%s\">%s</text>\n",startX, startY, fontSize, verticalAlignement,text));
-    }
-
-    private String generatePolarGraph(RailWayLW.TrainNoiseSource noiseSource, int heightIndex, double frequency, double minimumAttenuation, double maximumAttenuation) {
-        double dwidth = 500;
-        double dheight = 500;
-        double radius = dwidth/2 - 80;
-        double textRadius = radius + 25;
-        double centerx = dwidth/2;
-        double centery = dheight/2;
-        // HEADER
-        StringBuilder sb = new StringBuilder("<!DOCTYPE html>\n" +
-                "<html>\n" +
-                "<body>\n" +
-                "<svg height=\"500\" width=\"500\">\n");
-        // CIRCLES
-        sb.append(String.format(Locale.ROOT, "<circle cx=\"%f\" cy=\"%f\" r=\"%f\" stroke=\"black\" " +
-                "stroke-width=\"1\"  fill=\"transparent\"/>\n",centerx, centery, radius));
-        sb.append(String.format(Locale.ROOT, "<circle cx=\"%f\" cy=\"%f\" r=\"%f\" stroke=\"black\" " +
-                        "stroke-width=\"1\"  fill=\"transparent\"  stroke-dasharray=\"5,5\"/>\n",
-                centerx, centery, radius*0.75));
-        sb.append(String.format(Locale.ROOT, "<circle cx=\"%f\" cy=\"%f\" r=\"%f\" stroke=\"black\" " +
-                        "stroke-width=\"1\"  fill=\"transparent\"  stroke-dasharray=\"5,5\"/>\n",
-                centerx, centery, radius*0.5));
-        sb.append(String.format(Locale.ROOT, "<circle cx=\"%f\" cy=\"%f\" r=\"%f\" stroke=\"black\" " +
-                        "stroke-width=\"1\"  fill=\"transparent\"  stroke-dasharray=\"5,5\"/>\n",
-                centerx, centery, radius*0.25));
-
-        for(int angle=0; angle < 360; angle += 30) {
-            double destX = centerx + Math.cos((angle / 180.0) * Math.PI) * radius;
-            double destY = centery + Math.sin((angle / 180.0) * Math.PI) * radius;
-            // Reverse order and rotate by 90°
-            int adjustedAngle = (630 - angle) % 360;
-            if(angle % 90 != 0) {
-                // Dashed segment lines
-                generateDashedLine(sb, centerx, centery, destX, destY);
-            } else {
-                // Plain segment lines
-                generateLine(sb, centerx, centery, destX, destY);
-            }
-            double textX = centerx + Math.cos((angle / 180.0) * Math.PI) * textRadius;
-            double textY = centery + Math.sin((angle / 180.0) * Math.PI) * textRadius;
-            generateText(sb, textX, textY, 25, String.format(Locale.ROOT, "%d", adjustedAngle), "middle");
-        }
-        // Attenuation levels legend
-        generateText(sb, centerx, centery,10,String.format(Locale.ROOT,
-                "%.0f dB", minimumAttenuation), "Hanging");
-        generateText(sb, centerx + radius * 0.25, centery,10,String.format(Locale.ROOT,
-                "%.0f dB", minimumAttenuation + (maximumAttenuation - minimumAttenuation) * 0.25), "Hanging");
-        generateText(sb, centerx + radius * 0.5, centery, 10,String.format(Locale.ROOT,
-                "%.0f dB", minimumAttenuation + (maximumAttenuation - minimumAttenuation) * 0.5), "Hanging");
-        generateText(sb, centerx + radius * 0.75, centery, 10,String.format(Locale.ROOT,
-                "%.0f dB", minimumAttenuation + (maximumAttenuation - minimumAttenuation) * 0.75), "Hanging");
-        generateText(sb, centerx + radius, centery, 10,String.format(Locale.ROOT,
-                "%.0f dB", minimumAttenuation + (maximumAttenuation - minimumAttenuation)), "Hanging");
-        // Generate attenuation curve
-        StringBuilder path = new StringBuilder();
-        for(int angle=0; angle < 360; angle += 1) {
-            int adjustedAngle = (630 - angle) % 360;
-            double phi = (adjustedAngle / 180.0) * Math.PI;
-            double theta = Math.PI / 2.0;
-            double attenuation = RailWayLW.getDirectionAttenuation(noiseSource, heightIndex, phi, theta, frequency);
-            double maxLevelX = centerx + Math.cos((angle / 180.0) * Math.PI) * radius;
-            double maxLevelY = centery + Math.sin((angle / 180.0) * Math.PI) * radius;
-            double attenuationPercentage = (attenuation - minimumAttenuation) / (maximumAttenuation - minimumAttenuation);
-            Vector2D interpolatedVector = Vector2D.create(new Coordinate(centerx, centery), new Coordinate(maxLevelX, maxLevelY));
-            interpolatedVector = interpolatedVector.multiply(attenuationPercentage);
-            double stopX = centerx + interpolatedVector.getX();
-            double stopY = centery + interpolatedVector.getY();
-            if(angle == 0) {
-                path.append(String.format(Locale.ROOT, "M %.1f %.1f", stopX, stopY));
-            } else {
-                path.append(String.format(Locale.ROOT, " L %.1f %.1f", stopX, stopY));
-            }
-            if(angle % 30 == 0) {
-                path.append("\n");
-            }
-        }
-        path.append(" Z");
-        sb.append("<path d=\"");
-        sb.append(path.toString());
-        sb.append("\"  fill=\"transparent\" stroke=\"red\" stroke-width=\"2\"/>\n");
-        //  <line x1="250" y1="250" x2="450" y2="250" stroke="black" stroke-width="1" />
-        sb.append("</svg> \n" +
-                "</body>\n" +
-                "</html>\n ");
-        return sb.toString();
-    }
-
-    @Test
-    public void exportDirectivity() throws IOException {
-        try(BufferedWriter bw = new BufferedWriter(new FileWriter("target/test.html"))) {
-            bw.write(generatePolarGraph(RailWayLW.TrainNoiseSource.ROLLING, 1, 500, -25, 0));
-        }
-    }
-
    /* @Test
     public void Test_Cnossos_Rail_emission_section_6() {
         String vehCat = "SNCF-BB66400";
